Initial NewUI commit
This commit contains all changes made for the NewUI branch up to NewUI Preview 2. It took a while to get this branch into a clean state, so there are a lot of changes here.

Major additions:

- New ImGui style
- Add Overlay Bar
- Add Floating Settings & Overlay Properties window with partial settings reimplementations
- Add VR Keyboard
- Add TranslationManager & translation support
- Add window scroll-pull overlay creation for desktop overlays
- Add new drag creation for desktop and Performance Monitor overlays from OverlayBar
- Add "Add focused Window as Overlay" button to Floating UI
- Add AuxUI
- Add auto docking
- Upgrade WindowManager to manage capturable window list and update it automatically
- Rearrange Floating UI to be attached on the center bottom of overlays, with the Action-Bar appearing in the same spot and the main buttons sliding out of the way when visible
- Automatically restore window captures when a previously used window appears again
- Add setting to control On Capture Lost behavior (Nothing/Hide Overlay/Remove Overlay)
- Add setting to enable strict window matching

Major removals:

- Remove Keyboard Helper
- Rework overlay texture space setup
- Remove picker support from WinRT library
- Remove -2 Desktop Duplication default value + handling
- Remove UI SteamVR keyboard input support

Minor additions:

- Add animated drop-down widgets
- Add new string replace utility functions
- Add TextColoredUnformatted()
- Add ComputeOverlayIntersectionForDevice() utility function
- Add support for HmdVector3_t to Vector3
- Add FloatingUI "window add" icon
- Add FindPointerDeviceForOverlay() utility function
- Store class name for window overlays
- Keep last window state updated on dashboard app as well
- No longer set WindowManager inactive during runtime and track overlay active state instead to reduce amount of window tracking when not needed
- Use class name in window matching algorithm
- Store last known HICON in overlay config and use it when the window is no longer available

Minor changes & fixes:

- Fix DPRect formatting and warnings
- Fix deadlock in SliderWithButtonsFloatPercentage()
- Fix SliderWithButtons*() not forwarding ItemDeactivated state properly
- Fix race condition in DPWinRT_Init()
- Fix DetachedTransformSyncAll() not restoring current overlay state
- Fix ImGui::StringContainsUnmappedCharacter() crash when no font was loaded yet
- Fix ImGui::StringContainsUnmappedCharacter() accessing invalid font pointer when called outside of ImGui frame
- Fix ImGui mouse button state not release when the overlay or its input setting disappears
- Fix jumplists and Action Center being detected as capturable windows
- Clamp mouse coordinates to the hovered overlay texture space
- Adjust variable names
- Rename OverlayManager::AddOverlay() to DuplicateOverlay()
- Optimize ImGui_ImplOpenVR_SetIntersectionMaskFromWindows() to take an array of overlays to apply to at once
- Move drag-end OutputManager code to OnDragFinish()
- Implement additional overlay position offset
- Rework overlay transform resetting
- Adjust dashboard origin transform origin
- Remove configid_bool_overlay_detached
- Rename configid_int_overlay_detached_origin to configid_int_overlay_origin
- Remove backwards compatibility config quirks
- Remove dashboard overlay specialization
- Remove dashboard overlay special handling for ID 0 & other legacy handling code
- Minimize dashboard origin Y-position fluctuations from HMD pose angle
- Convert transform on overlay origin change
- Only use single transform in OverlayConfigData
- Use Meiryo instead of MS Gothic for Japanese characters if possible
- Replace overlay transform string sync with async float syncing
- Use different config file name for the NewUI preview branch
- Move OffsetTransformFromSelf() into Matrix4 class & rename HmdMatrix34_t OffsetTransformFromSelf() to TransformOpenVR34TranslateRelative()
- Improve stability of dashboard overlay origin
- Align Performance Monitor window to the center of the overlay so temp drags work at all sizes
- Fix FPS being grayed out in compact style when no scene app in running
- Rename WindowMainBar to WindowFloatingUIActionBar & WindowSideBar to WindowFloatingUIMainBar
- Merge WindowMainBar and WindowSideBar source files
- Rename WindowMainBar.h/cpp to WindowFloatingUIBar.h/cpp
- Don't reload textures while a widget is active or has just be released
- Rename UIManager::IsOverlayVisible() to IsOverlayBarOverlayVisible()
- Remove DashboardUI & old WindowSettings files
- Reset last window state on explicit null HWND messages
- Only increase Floating UI sort order when it's being hovered
- Repeat frame when setting texture reload flag
- Sync overlay HWND in response to ipcact_sync_config_state if a window capture is active
- Updated default config
- Code cleanup
- Bump version to NewUI Preview 2 (2.9.9.2)

+++ src/Shared/DPRect.h
@@ -12,11 +12,11 @@ public:
     Vector2Int      Min;    // Upper-left
     Vector2Int      Max;    // Lower-right
 
-    DPRect()                                             : Min(0.0f, 0.0f), Max(0.0f, 0.0f) {}
+    DPRect()                                             : Min(0, 0), Max(0, 0)             {}
     DPRect(const Vector2Int& min, const Vector2Int& max) : Min(min), Max(max)               {}
     DPRect(int x1, int y1, int x2, int y2)               : Min(x1, y1), Max(x2, y2)         {}
 
-    Vector2Int  GetCenter() const                   { return Vector2Int((Min.x + Max.x) * 0.5f, (Min.y + Max.y) * 0.5f); }
+    Vector2Int  GetCenter() const                   { return Vector2Int(int((Min.x + Max.x) * 0.5f), int((Min.y + Max.y) * 0.5f)); }
     Vector2Int  GetSize() const                     { return Vector2Int(Max.x - Min.x, Max.y - Min.y); }
     int         GetWidth() const                    { return Max.x - Min.x; }
     int         GetHeight() const                   { return Max.y - Min.y; }
@@ -27,9 +27,9 @@ public:
     bool        Contains(const Vector2Int& p) const { return p.x     >= Min.x && p.y     >= Min.y && p.x     <  Max.x && p.y     <  Max.y; }
     bool        Contains(const DPRect& r) const     { return r.Min.x >= Min.x && r.Min.y >= Min.y && r.Max.x <= Max.x && r.Max.y <= Max.y; }
     bool        Overlaps(const DPRect& r) const     { return r.Min.y <  Max.y && r.Max.y >  Min.y && r.Min.x <  Max.x && r.Max.x >  Min.x; }
-    void        Add(const Vector2Int& p)            { if (Min.x > p.x)     Min.x = p.x;     if (Min.y > p.y)     Min.y = p.y;     if (Max.x < p.x)     Max.x = p.x;     if (Max.y < p.y)     Max.y = p.y; }
+    void        Add(const Vector2Int& p)            { if (Min.x > p.x)     Min.x = p.x;     if (Min.y > p.y)     Min.y = p.y;     if (Max.x < p.x)     Max.x = p.x;     if (Max.y < p.y)     Max.y = p.y;     }
     void        Add(const DPRect& r)                { if (Min.x > r.Min.x) Min.x = r.Min.x; if (Min.y > r.Min.y) Min.y = r.Min.y; if (Max.x < r.Max.x) Max.x = r.Max.x; if (Max.y < r.Max.y) Max.y = r.Max.y; }
-    void        Expand(const int amount)            { Min.x -= amount;   Min.y -= amount;   Max.x += amount;   Max.y += amount; }
+    void        Expand(const int amount)            { Min.x -= amount;   Min.y -= amount;   Max.x += amount;   Max.y += amount;   }
     void        Expand(const Vector2Int& amount)    { Min.x -= amount.x; Min.y -= amount.y; Max.x += amount.x; Max.y += amount.y; }
     void        Translate(const Vector2Int& d)      { Min.x += d.x; Min.y += d.y; Max.x += d.x; Max.y += d.y; }
     void        TranslateX(int dx)                  { Min.x += dx; Max.x += dx; }
